# internal/docs/swagger.yaml
basePath: /
definitions:
  internal_account.accountResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  internal_account.loginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  internal_account.loginResponse:
    properties:
      account_id:
        type: integer
      refresh_token:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  internal_account.registerRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  internal_feed.FeedAuthor:
    properties:
      id:
        type: integer
      username:
        type: string
    type: object
  internal_feed.FeedVideoItem:
    properties:
      author:
        $ref: '#/definitions/internal_feed.FeedAuthor'
      cover_url:
        type: string
      create_time:
        type: integer
      description:
        type: string
      id:
        type: integer
      is_liked:
        type: boolean
      likes_count:
        type: integer
      play_url:
        type: string
      title:
        type: string
    type: object
  internal_feed.ListByFollowingRequest:
    properties:
      latest_time:
        type: integer
      limit:
        type: integer
    type: object
  internal_feed.ListByFollowingResponse:
    properties:
      has_more:
        type: boolean
      next_time:
        type: integer
      video_list:
        items:
          $ref: '#/definitions/internal_feed.FeedVideoItem'
        type: array
    type: object
  internal_feed.ListByPopularityRequest:
    properties:
      as_of:
        type: integer
      latest_before:
        type: string
      latest_id_before:
        type: integer
      latest_popularity:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
    type: object
  internal_feed.ListByPopularityResponse:
    properties:
      as_of:
        type: integer
      has_more:
        type: boolean
      next_latest_before:
        type: string
      next_latest_id_before:
        type: integer
      next_latest_popularity:
        type: integer
      next_offset:
        type: integer
      video_list:
        items:
          $ref: '#/definitions/internal_feed.FeedVideoItem'
        type: array
    type: object
  internal_feed.ListLatestRequest:
    properties:
      latest_time:
        type: integer
      limit:
        type: integer
    type: object
  internal_feed.ListLatestResponse:
    properties:
      has_more:
        type: boolean
      next_time:
        type: integer
      video_list:
        items:
          $ref: '#/definitions/internal_feed.FeedVideoItem'
        type: array
    type: object
  internal_feed.ListLikesCountRequest:
    properties:
      id_before:
        type: integer
      likes_count_before:
        type: integer
      limit:
        type: integer
    type: object
  internal_feed.ListLikesCountResponse:
    properties:
      has_more:
        type: boolean
      next_id_before:
        type: integer
      next_likes_count_before:
        type: integer
      video_list:
        items:
          $ref: '#/definitions/internal_feed.FeedVideoItem'
        type: array
    type: object
...
host: 127.0.0.1:8080
info:
  contact: {}
  description: 短视频社区后端 API:账号、视频、Feed、互动、通知。
  title: Video Feed API
  version: "1.0"
paths:
  /account/login:
    post:
      consumes:
      - application/json
      description: 用户名密码登录,返回 access token 与 refresh token
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_account.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_account.loginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - account
  /account/register:
    post:
      consumes:
      - application/json
      description: 根据用户名和密码创建新账号
      parameters:
      - description: 注册参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_account.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_account.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: username taken
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - account
...
  /feed/listByFollowing:
    post:
      consumes:
      - application/json
      description: 仅返回当前登录用户关注的作者发布的视频,时间游标分页
      parameters:
      - description: 分页参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_feed.ListByFollowingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feed.ListByFollowingResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 关注流
      tags:
      - feed
  /feed/listByPopularity:
    post:
      consumes:
      - application/json
      description: 基于 Redis 时间窗口分片 ZSET 的热度榜单,as_of 快照分页
      parameters:
      - description: 分页参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_feed.ListByPopularityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feed.ListByPopularityResponse'
      summary: 热榜
      tags:
      - feed
  /feed/listByTag:
    post:
      consumes:
      - application/json
      description: 根据标签名筛选视频列表
      parameters:
      - description: 标签筛选参数
        in: body
        name: body
        required: true
        schema:
          properties:
            limit:
              type: integer
            tag_name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 按标签筛选
      tags:
      - feed
  /feed/listLatest:
    post:
      consumes:
      - application/json
      description: 按发布时间倒序的最新 Feed,游标分页(time + id 复合)
      parameters:
      - description: 分页参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_feed.ListLatestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feed.ListLatestResponse'
      summary: 最新视频流
      tags:
      - feed
  /feed/listLikesCount:
    post:
      consumes:
      - application/json
      description: 按点赞数倒序的视频榜单,复合游标 (likes_count + id) 避免相同点赞数排序漂移
      parameters:
      - description: 分页参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_feed.ListLikesCountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_feed.ListLikesCountResponse'
      summary: 点赞榜
      tags:
      - feed
...
securityDefinitions:
  BearerAuth:
    description: Bearer JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"